- Improve per-object motion blur a bit. - Add OnFrameStart and OnFrameEnd to components. Currently, TransformComponent uses OnFrameEnd to cache the previous frame's world matrix. - Add some editor controls to AnimationComponent. - Fix skeleton rendering for animated nodes. - Scale down grass on slopes. - Add LoD variants to the grass system (the lower the lod the less billboards used). - Try and fix gamma correction and the weird tonemapping issue in the editor when switching debug views. - Fix the "infinity plane". - Fix prepass shader generation in Material.cpp for nodes that need to output velocity.

+++ assets/shaders/GLSL/blur.glsl
@@ -272,12 +272,12 @@ uniform int dvd_maxSamples;
 out vec4 _outColour;
 
 void main(void) {
-    const vec2 texelSize = 1.0f / vec2(textureSize(texScreen, 0));
+    const vec2 texelSize = 1.f / vec2(textureSize(texScreen, 0));
     const vec2 screenTexCoords = gl_FragCoord.xy * texelSize;
     const vec2 velocity = texture(texVelocity, screenTexCoords).rg * dvd_velocityScale;
 
-    const int speed = int(length(velocity / texelSize));
-    const int nSamples = clamp(speed, 1, dvd_maxSamples);
+    const float speed = length(velocity / texelSize);
+    const int nSamples = clamp(int(speed), 1, dvd_maxSamples);
 
     _outColour = texture(texScreen, screenTexCoords);
